Branch Compiler Errors Fixed

+++ Assets/Source/Actions/Systems/ActionAttributeManager.cs
@@ -13,14 +13,14 @@ namespace Action
         public void CreateActionPropertyType(Contexts contexts, Enums.ActionType actionType, string name)
         {
             ActionPropertyEntity = contexts.actionProperties.CreateEntity();
-            ActionPropertyEntity.AddActionProperty((int)actionType);
+            ActionPropertyEntity.AddActionProperty(actionType);
             ActionPropertyEntity.AddActionPropertyName(name);
         }
 
         public void CreateActionPropertyType(Contexts contexts, Enums.ActionType actionType)
         {
             ActionPropertyEntity = contexts.actionProperties.CreateEntity();
-            ActionPropertyEntity.AddActionProperty((int)actionType);
+            ActionPropertyEntity.AddActionProperty(actionType);
             ActionPropertyEntity.AddActionPropertyName(actionType.ToString());
         }
